:twisted_rightwards_arrows: Merge(feature/be/gameAPI리팩토링-> develop): 게임 안 끝나는 오류수정

See merge request s09-final/S09P31A501!156

Former-commit-id: ad6def29f9d29d8222423b9ca9df432c704cb9d6

+++ backend/src/main/java/com/npc/say_vr/domain/game/domain/Ranking.java
@@ -33,7 +33,7 @@ public class Ranking {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private Long point;
+    private int point;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "tier_id")
