+++ backend/src/main/java/com/npc/say_vr/domain/game/domain/Tier.java
@@ -26,5 +26,5 @@ public class Tier {
     private String name;
     private String image;
     private int minPoint;
-    private Long maxPoint;
+    private int maxPoint;
 }
