+++ backend/src/main/java/com/npc/say_vr/domain/game/repository/RankingRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface RankingRepository extends JpaRepository<Ranking, Long> {
 
     Optional<Ranking> findByUserId(Long userId);
-    Long countByPointGreaterThan(Long point);
+    Long countByPointGreaterThan(int point);
 
 }
